refactor(auth): restructure auth and redis configurations
- Move rate-limiting and Redis configurations into dedicated files.
- Update imports to reflect new structure.
- Refactor rate limit logic for maintainability.
- Include Redis health checks in the API to ensure connectivity.
- Adjust pnpm dependencies for new imports.
