+++ apps/api/package.json
@@ -27,6 +27,8 @@
 		"gateway": "workspace:*",
 		"hono": "^4.7.6",
 		"hono-openapi": "^0.4.6",
+		"ioredis": "5.7.0",
+		"nodemailer": "^7.0.5",
 		"openai": "^5.19.1",
 		"posthog-node": "4.18.0",
 		"stripe": "18.1.1",
@@ -34,6 +36,7 @@
 		"zod-openapi": "^4.2.4"
 	},
 	"devDependencies": {
-		"@types/node": "^24.0.3"
+		"@types/node": "^24.0.3",
+		"@types/nodemailer": "^6.4.17"
 	}
 }
