+++ apps/api/src/auth/handler.ts
@@ -1,5 +1,6 @@
 import { OpenAPIHono } from "@hono/zod-openapi";
-import { auth } from "@llmgateway/auth";
+
+import { apiAuth } from "./config";
 
 import type { ServerTypes } from "../vars";
 
@@ -7,7 +8,7 @@ import type { ServerTypes } from "../vars";
 export const authHandler = new OpenAPIHono<ServerTypes>();
 
 authHandler.use("*", async (c, next) => {
-	const session = await auth.api.getSession({ headers: c.req.raw.headers });
+	const session = await apiAuth.api.getSession({ headers: c.req.raw.headers });
 
 	if (!session) {
 		c.set("user", null);
@@ -21,5 +22,5 @@ authHandler.use("*", async (c, next) => {
 });
 
 authHandler.on(["POST", "GET"], "/auth/*", (c) => {
-	return auth.handler(c.req.raw);
+	return apiAuth.handler(c.req.raw);
 });
