+++ apps/api/src/vars.ts
@@ -1,4 +1,4 @@
-import type { Variables } from "@llmgateway/auth";
+import type { Variables } from "./auth/config";
 import type { Env } from "hono/types";
 
 export interface ServerTypes extends Env {
