+++ packages/auth/package.json
@@ -18,10 +18,6 @@
 		"build": "tsc && tsup"
 	},
 	"dependencies": {
-		"@llmgateway/logger": "workspace:*",
-		"@types/nodemailer": "^6.4.14",
-		"better-auth": "1.2.7",
-		"ioredis": "^5.4.1",
-		"nodemailer": "^7.0.5"
+		"better-auth": "1.2.7"
 	}
 }
