+++ packages/auth/src/index.ts
@@ -1,2 +1,1 @@
 export * from "./auth";
-export * from "./rate-limit";
